smartmealplanner/smartmealplanner/requirements.txt
Django
django-cors-headers
djangorestframework
d psycopg2-binary
